Move code for updating sim state into `simulation::update`
